Update for Lib. v2022.1 - Readme now shows right no. of IRQs
The readme file:./Libero_Projects/README.md was showing the wrong
number of 'Internal IRQs' for the 3 configurations available for
MIV_RV32. The number was changed from 6 to 1. The designs used
to support 6 'Internal IRQs' up until v12.6 release of the scripts,
when that number was reduced to 1 as, no more interrupts were needed.

Remainder of the DGC2 low TCL scripts have been renamed
and adjusted for new format.

Author: Sebastian Slowikowski <seb.slowikowski@microchip.com>
Signed-off: Sebastian Slowikowski <seb.slowikowski@microchip.com>